pyproject.toml
[project]
name = "agenticlanggraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-groq>=0.3.4",
    "langchain-tavily>=0.2.5",
    "langgraph>=0.5.0",
    "langgraph-cli[inmem]>=0.3.3",
    "langsmith>=0.4.4",
    "python-dotenv>=1.1.1",
]
